pyproject.toml
[project]
name = "ppt-to-web"
version = "0.1.0"
description = "Convert PowerPoint presentations to professional web pages with McKinsey/Bloomberg/Economist style"
readme = "README.md"
authors = [
    { name = "Henry Yao", email = "henryyao1126@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
ppt-to-web = "ppt_to_web.cli:cli"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
